Add files via upload
A seasonality data of the ARIMA model returns an ARIMA (1,1,1). There is one significant lag in the autoregressive part of the model. The data needed to be differenced once to achieve stationarity. There is one significant lags in the moving average part of the model.The relatively low AIC and BIC values suggest that this model is a good fit for the data